Fix Storage file creation
