+++ src/main/java/miki/task/ToDo.java
@@ -6,6 +6,11 @@ public class ToDo extends Task {
     }
 
     @Override
+    public String toStorageFormat() {
+        return "T | " + super.toStorageFormat();
+    }
+
+    @Override
     public String toString() {
         return "[T] " + super.toString() + "\n";
     }
